Fix translations scripts

+++ pipeline/translate/translate-mono.sh
@@ -8,6 +8,7 @@ set -euo pipefail
 
 test -v GPUS
 test -v MARIAN
+test -v WORKSPACE
 
 
 mono_path=$1
@@ -17,11 +18,12 @@ output_path=$3
 config=${model_dir}/model.npz.best-ce-mean-words.npz.decoder.yml
 decoder_config=${WORKDIR}/pipeline/translate/decoder.yml
 tmp_dir=$(dirname $output_path)/tmp
+
 mkdir -p $tmp_dir
 
 
 # Split the corpus into smaller chunks.
-test -s $tmp_dir/file.00 || pigz -dc $mono_path | split -d -l 2000000 - $tmp_dir/file.
+test -s $tmp_dir/file.00 || pigz -dc $mono_path | split -d -l 500000 - $tmp_dir/file.
 
 # Translate source sentences with Marian.
 # This can be parallelized across several GPU machines.
@@ -37,9 +39,12 @@ cat $tmp_dir/file.??.out | pigz > $output_path
 
 # Source and artificial target files must have the same number of sentences,
 # otherwise collect the data manually.
-echo "# sentences $mono_path vs $output_path"
-pigz -dc $mono_path | wc -l
-pigz -dc $output_path | wc -l
+src_len=$(pigz -dc $mono_path | wc -l)
+trg_len=$(pigz -dc $output_path | wc -l)
+if [[ src_len != trg_len ]]; then
+    echo "Error: length of ${mono_path} ${src_len} is different from ${output_path} ${trg_len}"
+    exit 1
+fi
 
 rm -rf $tmp_dir
